requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
requests==2.32.3
datasets==2.19.1
ragas==0.1.16
pinecone-client==5.0.1